refactor: begin refactor to sqlite

+++ src/main/java/com/ceuma/connectfono/repositories/DependentRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.UUID;
 
-public interface DependentRepository extends JpaRepository<Dependent, Integer> {
+public interface DependentRepository extends JpaRepository<Dependent, Long> {
     @Query(value = "SELECT * FROM dependent WHERE patient_id = :patient_id", nativeQuery = true )
-    List<Dependent> getByPatientId(@Param("patient_id") Integer patientId);
+    List<Dependent> getByPatientId(@Param("patient_id") Long patientId);
 }
