+++ src/main/java/com/ceuma/connectfono/repositories/PatientRepository.java
@@ -7,9 +7,8 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
+public interface PatientRepository extends JpaRepository<Patient, Long> {
     @Query(value = "SELECT * FROM patient WHERE type = 'ALUNO'", nativeQuery = true)
     List<Patient> findAllAlunos();
