+++ src/main/java/com/ceuma/connectfono/repositories/QuestionsRepository.java
@@ -3,7 +3,5 @@ package com.ceuma.connectfono.repositories;
 import com.ceuma.connectfono.models.Questions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.UUID;
-
-public interface QuestionsRepository  extends JpaRepository<Questions, Integer> {
+public interface QuestionsRepository  extends JpaRepository<Questions, Long> {
 }
