+++ src/main/java/com/ceuma/connectfono/repositories/StaffRepository.java
@@ -6,9 +6,8 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
-import java.util.UUID;
 
-public interface StaffRepository extends JpaRepository <Staff, Integer> {
+public interface StaffRepository extends JpaRepository <Staff, Long> {
     @Query(value = "SELECT * from staff WHERE email = :email AND password = :password ", nativeQuery = true)
     Staff login(@Param("email")String email, @Param("password") String password);
